pyproject.toml
[project]
name = "cyborg-fu"
version = "2.0.0"
description = "A pygame arcade fighting game - modernized from 2010 Python 2 original"
authors = [{name = "Stephen Sanchez"}]
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.3",
]

[project.scripts]
cyborg-fu = "cyborg_fu.__main__:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cyborg_fu*"]

[tool.setuptools.package-data]
cyborg_fu = ["../data/*.png"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = "pygame.*"
ignore_missing_imports = true

[tool.pylint.main]
py-version = "3.13"

[tool.pylint."messages control"]
disable = [
    "too-few-public-methods",
    "too-many-instance-attributes",
    "too-many-arguments",
    "too-many-locals",
    "too-many-positional-arguments",
    "duplicate-code",
]

[tool.pylint.format]
max-line-length = 100
